Fix "ClassFormatError : Truncated class file" introduced with use of JarJar
Symptoms :
On JBoss, the following error could occur when using jar built with jarjar.

[org.jboss.modules] (ServerService Thread Pool -- 75) Failed to define class org.mozilla_yui.javascript.Parser in Module "deployment.jcms10.war:main" from Service Module Loader: java.lang.ClassFormatError: Failed to link org/mozilla_yui/javascript/Parser (Module "deployment.jcms10.war:main" from Service Module Loader): Truncated class file

        at sun.reflect.NativeConstructorAccessorImpl.newInstance0(Native Method)
        at sun.reflect.NativeConstructorAccessorImpl.newInstance(NativeConstructorAccessorImpl.java:62)
        at sun.reflect.DelegatingConstructorAccessorImpl.newInstance(DelegatingConstructorAccessorImpl.java:45)
        at java.lang.reflect.Constructor.newInstance(Constructor.java:423)
        at org.jboss.modules.ModuleClassLoader.defineClass(ModuleClassLoader.java:446)

Explanation :
jar task (and jarjar too) do not prevent the SAME class from being added
twice in zip. Because the file set included both rhino and YUI version of
the rhino Parser, the class was present twice.

Fix :
Reintroduce use of a temporary directory in which all files are unzip (thus duplicte ovewriten)
Then invoke jarjar to create the jar and perform the package rewrite.